fix(log-embedder): rewrote the question for getting title (#7739)
Spaces does not always count has tokens it depends on what follow.

This change removes them at the end of our question.
We instead adds a \n, so it's clear for ChatGPT to get when the log start.

This fixes the token length addition between the template and logs next
lines.

This refactors computing of truncated_log_tokens_length to always have it
in sync with truncated_log in logging and future Sentry.

Change-Id: I75e30771ebd210184bc941aae5ed704c6cd910c7