chore(fake.env): Don't use deprecated settings names (#7851)
Rename param in fake.env to use their current name instead of their alias to be able to use the vscode debugger. We don't know why but the alias doesn't work on import of the config. Probably something messed up with this lib : vscode/extensions/ms-python.python-2023.14.0/pythonFiles/lib/python/debugpy/adapter

It's not a big deal to use the current name instead of alias since they are tested in this test: `test_legacy_env_sets`