fix(worker): ensure bucket cleanup is always executed and never fails (#7892)
This is required in case a CancelledError is raised after we removed
the last pull request of the bucket, but before we clean the "streams,"
key. Otherwise, "streams" can contain organization buckets that don't have
pull requests to proceed.

Also, do a cleanup if we discover a empty bucket later, this can occur when 
the CancelledError has been raised during the cleanup_org_bucket.

This solves the flakytest test_dedicated_multiple_processes

Change-Id: I37a14261043f13bd69c2d67eef7f893e89709b34