refactor(date): Change `Schedule` behavior to return "next_datetime" closer to the start/end (#7970)
Before this commit we would return add an additional second to the start of the schedule, and an additional minute to the end of it to for the `get_next_datetime`.
After this commit we do not add an additional second to the start of the schedule, and we add only an additional second to the end of the schedule. This will help having "faster" conditions nullification/validation.

Related MRGFY-2657