fix(eventlogs): don't compute total for pg backend (#7979)
The current value is wrong, this is not as efficient as Redis to get the
total as this backend supports filters.

Conversing this means doing the pg query twices (or one huge and
unoptimizable query)

This attribute is not used anywhere. It has been implemented just to
kept iso with the old API.

Change-Id: I867ff791a1a825c0fdcc8e058dd4487b5e84711c