chore(log-embedder): Remove error when failed step can't be found (#8011)
The last case where failed step can't be found is when the job timeout or fails after all steps succeed (I can't explain this behaviour of github). In this case, logs are not relevant but job annotations are. So as a general rule when no failed step can be found, we download annotations.

Fixes: MRGFY-2642
Fixes: ENGINE-3EY