fix(queue): allow reembarking in refresh even in cancelled state (#7983)
Currently, if the PR is unqueue and the action is in cancelled/failure state.
The user has to waits the action queue to go back to pending state,
before it can use refresh to re-embark the pull request.

This change makes it possible to refresh and re-embark the pull request
earlier in the process when action state is still cancelled or failure.

Fixes MRGFY-2672

Change-Id: I7f11b0062d346d50f4e116caaa6edcfcd44fb6a7