fix(queue): report an error when we stop retrying (#8076)
Currently, when we reach the maximum number of merge tries, we don't
inform the user or even dequeue the PR.

This reports an error message to the user and sets a workflow automation
state as failure.

When the unexpected merge error occurs, this also adds many logging
information to understand what happen and be able to give more
information to GitHub Support.

Fixes MRGFY-2701

Change-Id: I4ad71f10d373345e9fecdef7b2c883d07448abd4