fix(applications): change the way to limit application keys (#7930)
This previous version can be bypassed for a short time when multiple on
going requests are in flight.

This change uses a trigger to maintain an application key counter into
the GitHub account table and check the limit there to reject the
insertion in application table.

To do so, this adds a new library alembic_utils to automatically
generate migration for Postgres specific trigger/function/...

And a mechanism to convert custom Postgres function exception to Python
exception.

Depends-On: #7953
Fixes MRGFY-2654

Change-Id: I754fcef15680ef5d0b2afb6a48976c63a3fd6eba