fix(queue): automatically requeue already ready pull request (#8209)
When the pull request is unqueue due to CI failure, we require the user
to take a look at the CI. But if the speculative_checks=1 and
in place checks are allowed, we can requeue the pull request automatically
like we did for the strict mode.

This change does that.

Fixes MRGFY-891

Change-Id: I5fd8d6a7f979f8e8c118a6b9c1a3c2ecdcc18256

Depends-On: #8208