refactor: drop `python-dateutil` dependency requirement (#8359)
The library is barely maintained (last release in 2021) and we only use it for `relativedelta`, which can easily be done manually.
`freezegun` dependency still uses it, hence why it's not fully dropped, but this library is also barely maintained (last release in 2022) so we might also need to drop it for another.

Those change are required if we want to use python 3.12 before the owners of those packages finally decide to fix the bugs and make a release on pypi.