feat: Add rerun on flaky check run (#8260)
The flaky check rerun mechanism works like this :

In the context of requesting data from 'consolidated_data' regarding checks, when there are failed checks marked for rerun in the configuration ('conf'), and they belong to a type we currently handle (limited to GitHub Actions at the moment), we assess whether they need to be rerun. At this point, there are three possibilities for each check:

- It must be rerun.
- It hasn't to be rerun.
- We don't have the information yet (e.g., the 'workflow_job' corresponding to the checks has not been fully processed).

Subsequently, if there are checks that need to be rerun or for which we lack information, we mark them as 'pending' in 'consolidated_data.' As a result, the rules evaluation will classify them as 'pending' rather than as failures.

Finally, if we need to rerun certain checks, we make the API call to trigger the rerun. For those checks for which we don't have information yet, as soon as we have sufficient data, we send a 'refresh' command to reevaluate the PR.

Replace: #8204 and #8205 

Fixes: MRGFY-2645