fix(api): check if mergify is enabled on the repository (#8416)
Currently, the repository check is done in get_repository_context().

It works well for private repositories, but not for public repositories.

This change leverages another GitHub API endpoint to get the installation object from the requested repository instead of the requested repository owner, this ensure to return 403 when Mergify is not installed anymore on the requested repository.

Fixes MRGFY-2784

Change-Id: I96c1925f7a1da403a2dd17dd122c0355faadbea8