fix(date): fix bug in `get_next_datetime` calculation of DateTimeRange (#8447)
In the case where we have a DateTimeRange with a `day` as an `UncertainDatePart`, a range that included the full day (`00:00` to `23:59`) and we were inside that range, the calculated next datetime was wrong. It returned the first day of the month while it should return the first day of the next month.