feat: Add model CiIssue (#8464)
Add the new model CiIssue for CI monitoring and link jobs to CiIssue by their
Cosine similarity.
As sentry we compute a short-id that has a customizable prefix
and an increment. The increment is unique by repo.

WorkflowJobLogNeighbours is still computed but will be useless soon,
it's kept for now to not blow-up everything related to it.

Fixes: MRGFY-2728

Change-Id: Iaf1443d7d09310bd12597b83367c17e36e9516de