chore: remove asgi_testing.py (#8580)
The current system doesn't play well with --reload of uvicorn
as the create_all() is rerun each a file is saved

This change removes the models.create_all() hack and move the https_only
into a settings.

This allows to use the same asgi script for local test and production.
And fixes the issue with --reload of uvicorn

Change-Id: I0fe92be24bc07ebb173b83764b49e0339fe9cde5