refactor(queues): split BaseUnqueueReason and BaseQueueCancelReason (#8648)
The current code has a single class for everything which is named
"unqueue" but actually does not remove the PR from the queue for real.

There are actually one things that is common, is that something is
canceled. Which is why this changes rename the base class in BaseQueueCancelReason.

Now, we also add a subclass BaseUnqueueReason which is only used when
things are canceled and the PR is also removed definitely from the
queue.

This should make it clearer for attributes like `queue-dequeue-reason`
to be able to return only BaseUnqueueReason instances.

Change-Id: I47bef37e568ebf9f75862050574eda1603814987