ci: use another command to check lock file (#8768)
The generated diff is currently always out of date as wheels may
released after we got a dependabot update, or maybe our local cache is
outdated and does have all wheels.

Not having all wheels is not an issue (it's just waste of CI time as we
will build the package from source).

The intent of the job is to check dependencies described in
pyproject.toml and the lock are in sync always.

This change uses another poetry command to just detect this without the
missing wheels noise.

Change-Id: I9e4762696f91ad4b53c47fcdff4b3907cff59625