refactor(events): move future fk on child table (#8841)
We face the issue of not being able to leverage ON DELETE CASCADE for this associated table because the FK is on the parent (event) which is deleted. The FK need to be relocated to the "speculative_check_pull_request" table to do so.

First step is to create the future FK to transfer the current one to the child table.

Related to [MRGFY-2789](https://linear.app/mergify/issue/MRGFY-2789)