revert: "fix: KeyError when queue_rule vanish (#8892)" (#8904)
I don't see why we should reset the whole train when this occurs.
Maybe pull requests ahead are queued into an existing queue, and we should not drop these train cars.

We should only remove/readd the problematique from the queue.

Also, we must not add new stuff on unexpected changes datastructure; I am working on removing this data structure because when used, the GitHub summary is wrong and stays pending forever.

This reverts commit f3916bcd65bcc91a160aced67171b7508d25546c.

Change-Id: I4ac7ca4022d39fa1b80a7c26a3b4339fb0e2c9c5