feat: Extract data from logs with chatGPT and save them in db (#8919)
Ask chat GPT4 to extract various data from logs and give it to us in a json. Save thses data in DB. 

In the same time and to avoid making to much call to OpenAI, replace the call to get a title for ci issue with the one for extractting the data and using the "problem_type" value as the title. 

Fixes: MRGFY-2903


Change-Id: Id42b55220035b044389088de91e19ff1d8f07930