refactor(events): copy existing parent id to future fk on child table (#8786)
We face the issue of not being able to leverage ON DELETE CASCADE for this associated table because the FK is on the parent (event) which is deleted. The FK need to be relocated to the "speculative_check_pull_request" table to do so.

Second step is to move existing FK on the parent to the corresponding child.

Related to [MRGFY-2789](https://linear.app/mergify/issue/MRGFY-2789)
Depends-On: #8841