refactor(events): remove old fk and create new constraint (#8941)
We face the issue of not being able to leverage ON DELETE CASCADE for this associated table because the FK is on the parent (event) which is deleted. The FK need to be relocated to the "speculative_check_pull_request" table to do so.

Third and last step we rebuild the new FK.

Related to [MRGFY-2789](https://linear.app/mergify/issue/MRGFY-2789)
Depends-On: https://github.com/Mergifyio/engine/pull/8786