chore(sentry): hide unwanted subapp error (#9003)
A bug in starlette/fastapi makes error subapp unhanlded error raise a
weird exception except returning 500.

The solution proposed by fastapi discusions was:
https://github.com/Mergifyio/engine/pull/8967

But this solution breaks Sentry and ddtrace asgi middlewares.

Their is no easy solution to the issue. So just hide the error in
Sentry.

The user will continue to  502 instead of 500, but well it's a server
side issue we need to fix. We already have the error in another Sentry thanks to
our logging middlewares, so it's not a big deal.

Fixes ENGINE-3JM

Change-Id: I1e95aa3249ac3d5dd1ee82309b3a29a4ca20f148