fix: do not insert or update pull_request commits if any of it can't be validated by pydantic (#9037)
We're seeing data such as:
```
commit: {
  author: None, 
  comment_count: 0, 
  committer: {
    date: '2023-11-28T09:15:11Z', 
    email: '', 
    name: ''
  }, 
...
```

Being returned by GitHub when requesting the list of commits. `committer` should be a complete `GitHubAccount` according to their schema/typing, but in this case, it misses some keys, and some values are empty.

Since there might be little value in storing those commits, we stick to the schema and ignore those commits entirely.

Fixes ENGINE-3PH