fix: do not insert pull request commits if the `/commits` endpoint returns a 404 (#9047)
Sometimes when querying commits of a pull request, we receive an HTTP error 404. In the engine we already have code to ignore some events if the `/pulls` endpoint returns an HTTP error 404, which is why this error is only happening now for this endpoint in the `github-in-postgres` worker.

This commit also add a test to check if a pull request, and its commits, from a fork are properly inserted inside our database.

Fixes ENGINE-3PJ