fix: use `exceptions.should_be_ignored` func when querying pull request commits (#9053)
When querying commits of a pull request in `github-in-postgres` service, we might receive exceptions we are alredy ignoring on other parts of the engine. We can use the `exceptions.should_be_ignored` function in order to do the same in this case in order to not have to manually handle every possible exceptions.

Fixes ENGINE-3PK