fix(events): migrate actor to JSON column (#9051)
GithubAuthenticatedActor can be either an application or a user. But the
primary key has been set on the ID column, while it should have been set
on the type + ID columns. We could fix the primary key, but this table
mixes two kinds of resources that can evolve differently (application IDs
will soon become UUIDs).

This PR replaces GithubAuthenticatedActor table by JSON columns. This
way, even if an actor has to change, events won't be affected by that
change.

Related to MRGFY-2965

Change-Id: Ifdebf28fd899f1d8491da4e4233c252acaeb1902