test: remove fixture `setup_new_event_loop` that does nothing (#9071)
This fixture appears to not be doing anything, none of our library defines it, yet when we remove it some unit tests breaks (the tests in `mergify_engine/tests/unit/merge_train/test_convoy.py`).
The fixture being moved (`mock_signals`) is to fix those unit tests breaking when the useless fixture is removed.