refactor: Split embedding and metadata extracting (#9030)
Split embedding and metadata extracting to avoid one way to impact the other.

Refactored `log_status` in `log_status`, `log_embedding_status`, and `log_metadata_extracting_status` to handle the error status separately for the 3 actions which are log downloading, log embedding, and extracting metadata from logs.

The refactor leaves ci_issue with no way to get a title from chatGPT. It's not a big deal since there is already a fallback in the API for this case. 

We assume that there is always a gcp client and that logs are potentially already downloaded from github and saved in  GCP. It allows us to download fewer data since in GCP we saved only the failed step log (at least for one of the embedding or extracting ways) and it reduces the number of requests to github.

The current way that the `def extract_metdata_way` is querying the db, will result in an extracting of metadata for all the WorkflowJob currently in db even the old one where the embedding is already computed.
Pro: We will have more data to compare the two ways and a big test for the embedding way
Con: It will  make a lot of call to OpenAI ChatGPT ($$$)

Fixes: MRGFY-2959

This query will be run on prod db after this PR is merged.
```
UPDATE gha_workflow_job 
        SET log_embedding_status = 'EMBEDDED'
        FROM (
            SELECT id
            FROM gha_workflow_job 
            WHERE log_embedding_status = 'UNKNOWN' and log_embedding is not NULL and log_status = 'DOWNLOADED'
            LIMIT 1000
        ) AS subquery
        WHERE gha_workflow_job.id = subquery.id;
```

Change-Id: I6c7e03334a6627418ff3fb09441340ae87a396ed