refactor(log_embedder): implement more separation of concerns (#9195)
This moves download code path to the WorkflowJob class itself, so it can
manages to retrieve the download of logs or annotation — and that's it.

The rest of the flow is still handled in the log embedder.

This as the upside to not duplicate the client instanciation and makes
it shareable if needs be at some point.

Change-Id: I6333f3f446af98fcc0329e2aebfea3e4f5e5a6de