feat: Add ci_issue_gpt created with data extracted from GPT (#9105)
Use metadata extracted from logs by chat gpt to create ci issues. Currently, the accuracy of ci issues gpt is not wonderful but it's not the aim of this PR. We will have to work on prompt or training gpt to get better metadata and on `CiIssueGPT.link_job_to_ci_issues` to get better use of these metadata.

Fixes: MRGFY-2906

Change-Id: Ic8bdb51b9677c3286a95db1aaa02d38fa50b497c