refactor(log_embedder): make embed_log use a log content (#9256)
The idea is to avoid splitting + reconstructing the log based on
splitlines + join all the time and do it only as necessary.

The test are also fixed to not include any \n as there is no \n sent to
the log cleaner and embedder, since splitlines is used to split raw
content.

Change-Id: I71d5fcedb0160523869df9daa6427762eeb30a4e