fix(queue): always set final reason of why a train car is deleted (#9186)
This change always sets the final conclusion for the Queue: check-run.

The reason is stored in TrainCarState.delete_reasons and serialized in
Redis and in the check-run.

Now get_dequeue_reason() will always get the right dequeue reason even
if the outcome of the train is not the reason the train car was deleted.

Fixes MRGFY-3005

Change-Id: If12472a558f4edda6adad78d99dcef233f2fef2a