fix(log-embedder): commit/rollback session correctly (#9332)
This change ensures that database error is raised only when we commit
the transaction and the session have been rollback correctly.

As, we have to do other update on same DB orm instance after SQL commit failed.

We same a untouched version of the instance before manipulating the one
in the session.

This ensures the next proceed job will not inherit a nonworking session object.

Fixes ENGINE-3Q5

Change-Id: I38222b4e00eecba8457c15e945bd24873534ac4c