refactor(queue): add retry and logging for merge-queue branch creation (#9250)
To correct a bug where a draft_pr branch is failing to be created, due to its ref potentialy already existing. The branch being deleted after receiving the exception, this leads to the PR being dequeued. We are now retrying the creation to see if we can create it properly after the first deletion, avoiding dequeuing the PR, in case it is simply due to a GitHub timing or some other external factors.

Fixes MRGFY-2955