test: remove `TestQueueActionWithPgEventsIngestion` test class (#9426)
Since we decided that we would have to refactor the workflow of the run engine, in order for it to be triggered by github events, after the data were updated in postgres, before enabling the `github-in-postgres` service for every clients (since GitHub events sending is sometimes really late, which create inconsistency in the merge train), those tests are useless.

They were initially added in order to make sure that the merge train was working properly even after pull requests, and related attributes, were selected from our Postgres database and before we reworked the workflow of the run engine, but with the decision specified above, they serve no purpose.
Even when we will refactor the workflow of the run engine, those tests wouldn't be relevant since the `github-in-postgres` service will be mandatory.

We already have tests to make sure that the pull requests, and its related attributes, are properly updated and selected from our postgres database, which is more than enough at the moment.