chore: change default check-run detail url (#9434)
In the past we was redirecting the detail url to itself as we don't have
a dashboard.

Setting the url to itself has as side effect that detail url differ when
we create the check-run and when we update it. This makes two API calls
for nothing for all check-run.

This changes the default details_url to the event log page of the dashboard.
To avoid this useless call.

Change-Id: Iba33fd329cdb9df0cfa875cf4b4c03b3b9aebf90