chore: replace unexpected changes (#9365)
This data structure was misleading after unexpected changes occurred
a PR may removed from the queue or the merge queue completely resetted.

Pausing the queue was also a reason to reset the queue, but was
classified as a train car outcome.

This change separates the different concerns.

Anything that reset the whole queue is now a "ResetEvent":

* merge-queue pause
* base branch pushed
* mergify support hard reset the queue

A new `BaseCancelReason` is associated ResetEvent called `MergeQueueReset`.

For DRAFT_PR_CHANGED and PULL_REQUEST_UPDATED, two new dequeue reasons have
been created:

* DRAFT_PULL_REQUEST_CHANGED
* PULL_REQUEST_UPDATED

Fixes MRGFY-3006

Change-Id: I09ae13ebd156ae338ee948fa63adbda57e58f900