fix(log-embedder): allow to configure used model (#9469)
The current model as only 5000000 TPD, out chat prompt + log is 16384 tokens.

So we can get the metadata of only ~300 logs per day

This change makes the model configurable, so we configure another one
in production until gpt-4-turbo is released and remove this limit.

I move the used model into the `ChatCompletio`n object as the model, response format and the code that read the response are linked together.

Change-Id: Ibd0b7909990c03273b804554e134e8aa557612d6