test: set `github-in-postgres` env variables only for tests that requires it (#9504)
Setting this for all the tests is useless since the majority of them do not run engine with the `github-in-postgres` service enabled.

This should also speed up a little bit the tests since now we have a few sql query less to execute before getting a pull request or its commits.