refactor: rework outcome computing (#9481)
This change introduces new outcomes and removes others:

MERGEABLE -> WAITING_FOR_MERGE
UNKNOWN has been split into:
* WAITING_FOR_CI -> CIs run
* WAITING_FOR_SCHEDULE -> CIs finish for schedule doesn't match yet
* WAITING_FOR_UNFREEZE -> CIs finish but queue is frozen

CHECKS_FAILED has been split into:
* CHECKS_FAILED -> CI failed and pull request inside is the culprit
* CHECKS_FAILED_LOOKING_FOR_ROOT_CAUSE -> checks failed but we are looking for the root cause

This allows to set checks_end_timestamps as soon as checks finish.

This simplifies code flow as the outcome is more precise, we don't have
to check everywhere why is it UNKNOWN or if CHECKS_FAILED is due to this
car or another ahead.

The reporting is also more precise as we can inform users why a pull
request checks are finished, but it has not been merged yet.

Related MRGFY-2696
Fixes MRGFY-3007

Change-Id: Ia44830335b1dcc79c54eb44e9c4b99e8a59a27bc