refactor(queue): use an exception to notify caller (#9533)
When we need to retry the merge, we have to pass a method to build a
PENDING summary and the caller is not really aware we are going to
retry or need to guess with the conclusion.

This change replace the mechanism by an exception so caller is aware
when this is required to retry the merge and can act in consequence.

Change-Id: Ie8c26334c6c28bb46ba59db2fffba695d5fc38ad