refactor(events): make event routing more extendable (#9534)
Currently, we need to build a class to route a event to a worker service, but if
an event needs to be routed to multiple places, we need one classes per
destinations combinaison. We have only three routes currently:
* stream
* github_in_postgres
* ci_monitioring

This change refactors it makes less cumbersome to route events to
services.

Change-Id: Ia3b4c4290922b321974da939c5b90e6c75967678