chore(gh-in-pg): ensure we never fail processing events (#9542)
Currently, we stop iterating the stream each time we trigger an
unexpected issue.

This changes ignore and log the error to continue.

This is not perfect if we reach
GITHUB_IN_POSTGRES_PROCESSING_BATCH_SIZE, we will need to introduce a
better retry mechanism at some points.

Since the default size is 1000, it should be good enough for now.

Change-Id: Iaa688f38a6434ad3a0843fd3a0f36935b59574ff