fix(worker): avoid running two identical workers in parallel (#9577)
Create a database lock for each engine worker to avoid running two identical worker in parallel. We just log the lock error for now. If this works as expected, we should raise an error or wait for the lock to be released.

We use [PostgreSQL Advisory Locks](https://www.postgresql.org/docs/current/functions-admin.html#FUNCTIONS-ADVISORY-LOCKS) to implement a global lock for each worker.

Fixes MRGFY-3011

Change-Id: I5ec4d951ffbc7c599b3fca2c475ac44c3e1149ef