refactor: split and simplify get_ci_issue (#9686)
The code for getting one issue and all issues is currently common.

Handling pagination/filtering for getting one issue is useless.

The change only uses ORM to load data, which make the query easier
to read.

This change should be more performant by design as it loads first the CiIssueGPT row and then loads its relationships.
Before we loaded the WorkflowJob table first with all its relationships and then kept only the workflow linked to the requested ci issue.

Change-Id: I5af137ec0c60778e53fd7e5d0662d3d3213377aa