ci: add pytest to import-checks when testing with dev deps (#9704)
Some import errors are not picked up with the checks we already have, but are picked up by pytest,
which is ran in tools/check-obsolete-fixtures.sh which leads to this script to fail and "hide"
the error by displaying all the folders as unused.
We still need to be able to run the import-checks with only prod deps.