refactor: simplify get_ci_issue_event_detail (#9703)
This change loads first the log_metadata requested row and then loads the
relationships.

This makes the code much more readable and more performance as we don't
have to read the whole WorkflowJob table.

Only the primary WorkflowJob primary key index is used.

The change only uses ORM to load data, which makes the query easier
to read.

Change-Id: I16d9d4501288efcc713e947f0fc6e6be36548fc4