refactor: drop EnhancedJob from ci-issues listing (#9694)
This uses the same method as the previous refactor (https://github.com/Mergifyio/engine/pull/9686), SQLAlchemy column_property to compute additional columns + selecting only useful columns.

This is to measure the performance of using subquery for
flaky/failed_run_count/pull_requests

On Mergifyio/docs, it timeout in production, SQL requests take more than
30s.

On Mergifyio/docs, in production with this branch, the same requests take 7s.

This is promising as the response payload returns a lot of useless data
to the frontend.

Change-Id: I73770edf22427dda3e992c868ee893116e2257a0