feat(ci_issue api): add pull_requests_impacted on events in ci_issue api detail (#9747)
Add pull_requests_impacted request on CiIssueEvent. New endpoint entry looks like this:
```
"pull_requests_impacted": [
    {
        "number": 123,
        "title": "fix: foobar",
        "author": "This guy",
        "events_count": 5,
    },
]
```

To compute those data I added a query_expression on `CiIssueGPT` that returned directly the expected JSON.

The list of pull_requests_impacted is sorted by:  events desc and number asc

For now, we don't limit the number of pull requests returned in this list so if a ci_issues is linked to 10000 PR it could be too long to load. If this is the case, we could easily add a limit parameter to `with_pull_requests_impacted_column` to get for example the 10 most impacted PR and have in front a "load more" button that would load the 10 next and so on with a pagination.

Fixes: MRGFY-3053

Change-Id: I2b33a60d6682b67d964eaecfc20db0b42684320b