chore(ci-issues): add new endpoint to list all events of an issue (#9842)
We have a listing page with all events that need to be paginated at some
point. The current design that list events inside a CI issue detail will
not allow that.

This change add the endpoint to get all events in a format that can be
paginated in the future.

Fixes MRGFY-3111

Change-Id: Ie4795217e3c0e84a04565891585090707d84b2cb