chore(log-embedder): Improving handling of missing file in zip (#9849)
It sometimes happens that a file is missing in the zip file when we download it, but it's there the next time. This PR is mainly here to avoid raising a sentry for nothing.

This PR also renames `UnexpectedLogEmbedderError` to `LogNotFoundInZipFileError` since it's only raised in this case.

Fixes MRGFY-3114
Fixes ENGINE-3SQ

Change-Id: I4c7ef8a70b3c9ac348140795209f077c9d6ad386