fix(queue): rework `get_outcome_from_conditions()` (#9949)
To check that we are waiting for a datetime condition, we check all the rule conditions without the datetime ones. The current implementation updates the datetime conditions to make them always pass. That doesn't work when we use the `not` operator:

```yaml
- not: --------------------------------------> False (reported as CONDITIONS_FAILED)
    or --------------------------------------> True
        - current-datetime = some-holiday ---> True
```

To make it work, we have to either ignore datetime conditions, not just make them pass, or use a TernaryFilter instead of a BinaryFilter.

Fixes MRGFY-3120

Change-Id: I8f6c4c21b79a0ef7997ccfa70a91687e9ee9baae