chore: track active_users with database (#10006)
The current data structure in Redis doesn't scale, it's main issue is
the method used to handle user/repository renaming that force a lot of
Redis key to be scanned.

Instead of complexify more this structure, this change starts to track
the usage in postgres.

In one month, we will be able to compare what we have in DB and in
Redis. And we can change the subscription usage API endpoint to use
postgres instead of Redis.

Related INC-125

Change-Id: I4d49ee4729d5cb697057e9e9132e388797274b24