fix(api-pulls): fix pagination of pulls api (#10042)
There were several issues with pagination on pulls api

The number of pr requested from GitHub API was not in sync with the number of pr requested from our api, leading to potentially fetching more data from GitHub than necessary.

- Sync them by sending the `per_page` parameter to GitHub API

The `nb_pulls_on_github` var that was used to build the cursor value was misplaced and was not taking into account pr that were skipped at the beginning of the loop.

- Syncing the number of pr returned by GitHub API removes the needs of this var and is replaced by the `pr_idx_on_page` in the cursor value.
  
When the last element fetched was the last element of the GitHub page we were returning an incorrect value for the cursor e.g. GitHub page 1 returned 5 results and we were sending cursor `1-4` that 
lead for the next call to fetch page 1 skipping all results then fetching page 2.

- Changing the way we compute the cursor at the end of the function fixes this

When we reached the last page, we returned a cursor with the value None as the next cursor which is equivalent to returning the first cursor, leading the caller of this API to loop over and over.

- Returning None instead of a cursor when we reach the last page fixes this